requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.5.2
decorator==4.3.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
inflection==0.3.1
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
-e git+git@github.com:wryfi/mntnr_hardware@180685b78bf29ba68864a818eb49bc61466050c7#egg=mntnr_hardware
-e git+git@github.com:wryfi/mountaineer@1df6f9a547643df1e05969f3d99f297abca429dd#egg=mountaineer
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyYAML==3.13
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.11
swagger-spec-validator==2.3.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
